backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase==2.4.5
httpx==0.26.0
python-jose[cryptography]==3.3.0
aioapns==3.1.0
apscheduler==3.10.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
feedparser==6.0.11
finnhub-python==2.4.19
openai==1.12.0
